dv_gen/requirements.txt
loguru==0.7.3
pandas==2.2.3
python-dotenv==1.0.1

